requirements.txt
PyQt5~=5.15.6
pyserial~=3.5
keyboard~=0.13.5